move autoguide the the main module
