+++ examples/neutra.py
@@ -26,13 +26,13 @@ import jax.numpy as np
 
 import numpyro
 from numpyro import optim
-from numpyro.contrib.autoguide import AutoBNAFNormal
 from numpyro.contrib.reparam import NeuTraReparam
 from numpyro.diagnostics import print_summary
 import numpyro.distributions as dist
 from numpyro.distributions import constraints
 from numpyro.distributions.util import logsumexp
 from numpyro.infer import ELBO, MCMC, NUTS, SVI
+from numpyro.infer.autoguide import AutoBNAFNormal
 
 
 class DualMoonDistribution(dist.Distribution):
