--- numpyro/contrib/nn/__init__.py
@@ -1,8 +1,0 @@
-# Copyright Contributors to the Pyro project.
-# SPDX-License-Identifier: Apache-2.0
-
-from numpyro.contrib.nn.auto_reg_nn import AutoregressiveNN
-from numpyro.contrib.nn.block_neural_arn import BlockNeuralAutoregressiveNN
-from numpyro.contrib.nn.masked_dense import MaskedDense
-
-__all__ = ['MaskedDense', 'AutoregressiveNN', 'BlockNeuralAutoregressiveNN']
