+++ numpyro/nn/__init__.py
@@ -1,0 +1,8 @@
+# Copyright Contributors to the Pyro project.
+# SPDX-License-Identifier: Apache-2.0
+
+from numpyro.nn.auto_reg_nn import AutoregressiveNN
+from numpyro.nn.block_neural_arn import BlockNeuralAutoregressiveNN
+from numpyro.nn.masked_dense import MaskedDense
+
+__all__ = ['MaskedDense', 'AutoregressiveNN', 'BlockNeuralAutoregressiveNN']
