+++ numpyro/nn/auto_reg_nn.py
