+++ test/test_flows.py
@@ -10,7 +10,7 @@ import pytest
 from jax import jacfwd, random
 from jax.experimental import stax
 
-from numpyro.contrib.nn import AutoregressiveNN, BlockNeuralAutoregressiveNN
+from numpyro.nn import AutoregressiveNN, BlockNeuralAutoregressiveNN
 from numpyro.distributions.flows import BlockNeuralAutoregressiveTransform, InverseAutoregressiveTransform
 from numpyro.distributions.util import matrix_to_tril_vec
